--- app/(authenticated)/admin/site-controller/components/PostStatus/PostStatus.tsx
@@ -1,9 +1,0 @@
-
-
-const PostStatus = () => {
-  return (
-    <div className="text-primary">PostStatus</div>
-  )
-}
-
-export default PostStatus
