--- app/(authenticated)/admin/site-controller/components/ProjectStatus/ProjectStatus.tsx
@@ -1,17 +1,0 @@
-import AddProjectDialouge from "./components/DiaologueBox/DialougeBox"
-
-
-
-const ProjectStatus = () => {
-  return (
-    <div className="text-primary">
-      ProjectStatus
-      <div className="mt-10">
-        <AddProjectDialouge/>
-      </div>
-    </div>
-
-  )
-}
-
-export default ProjectStatus
