+++ app/(authenticated)/admin/site-controller/post/PostStatus.tsx
@@ -1,0 +1,11 @@
+
+
+const PostStatus = () => {
+  return (
+    <div className="text-primary">
+      This is the Post Status Page
+    </div>
+  )
+}
+
+export default PostStatus
