+++ app/(authenticated)/admin/site-controller/project/ProjectStatus.tsx
@@ -1,0 +1,15 @@
+
+
+
+const ProjectStatus = () => {
+  return (
+    <section>
+      <div className="flex flex-col justify-center items-center">
+        
+
+      </div>
+    </section>
+  )
+}
+
+export default ProjectStatus
